feat(composer): split ds file into folder
