+++ crates/astria-composer/src/ds/mod.rs
@@ -1,0 +1,19 @@
+// Traits
+mod streaming_client;
+mod wire_format;
+
+// Provider Implementations for various rollups
+mod eth_provider;
+
+// Custom types
+mod types;
+
+// Wrappers
+mod sequencer_client;
+
+// Exports
+pub(crate) use self::sequencer_client::SequencerClient;
+pub(crate) use self::streaming_client::StreamingClient;
+pub(crate) use self::types::{RollupChainId, RollupTx, RollupTxExt};
+pub(crate) use self::wire_format::WireFormat;
+pub(crate) use self::eth_provider::EthProvider;
