+++ crates/astria-composer/src/ds/types.rs
@@ -1,0 +1,9 @@
+use ethers::types::Transaction as EthersTx;
+
+#[non_exhaustive]
+pub enum RollupTx {
+    EthersTx(EthersTx),
+}
+
+pub(crate) type RollupChainId = String;
+pub type RollupTxExt = (RollupTx, RollupChainId);
